bug fix for smooth_window (needs to be int)

+++ globin/input.py
@@ -495,7 +495,7 @@ class InputData(object):
 		# inversion algorithm specific parameters
 		self.ncycle = _find_value_by_key("ncycle", self.parameters_input, "default", 1, int)
 		window = _find_value_by_key("smooth_window", self.parameters_input, "default", "5", str)
-		self.gaussian_smooth_window = list(map(float, window.split(",")))
+		self.gaussian_smooth_window = list(map(int, window.split(",")))
 		std = _find_value_by_key("smooth_std", self.parameters_input, "default", "2.5", str)
 		self.gaussian_smooth_std = list(map(float, std.split(",")))
 		self.svd_tolerance = _find_value_by_key("svd_tolerance", self.parameters_input, "default", 1e-5, float)
